feat: enfore statefulset usage for app with disks instead of deployments
if there is more than 1 pod with disk, it can't work on multiple machines, only one (in theory). In practice, it starts multiple pods on the same physicial node. It doesn't brings any security at all and can let the end user think his application is deployed on multiple nodes and will always be available but not. Moving to statefulsetr is the only way to get volumes correctly managed