fix: stop relying on scw API for docker images operations
This CL changes how docker images are checked if exists and how they are
deleted.
Scaleway has a known bug / limitation on their end where the is a delay
between what's in the container registry and what their API sees.
While they are working on fixing that, we will using plain docker
command in the meantime allowing to speed up the engine not to have to
wait for scaleway API to be up-to-date.