fix: service to be created even if private port
As of today if port is set as private on an app / db, then no service
will be created for it in k8s. This should not be the case and a service
should be created even if app or db is flagged as private allowing other
internal apps to connect to it.