feat: aws kubeconfig retrieval improvment
This CL tries to introduce a better handling of S3 eventual consitency.
https://docs.aws.amazon.com/AmazonS3/latest/userguide/Welcome.html#ConsistencyModel

Issue was at cluster creation, we try to get a kubeconfig which doesn't
exists yet in order to know whether or not the cluster exists and needs
to be upgraded. It leads to S3 caching the fact this bucket doesn't
exists and reject later s# get while the bucket actually exists.

By introducing a retry mechanism, we allow s3 to properly propagate new
file / bucket creation and reach consistency.