feat: check helm q-storageclass install
This CL introduces a way to test Helm charts installation via a
`ChartInstallationChecker` trait on helm chart struct.
This checker is called after installation in order to check
whether or not the chart has been properly installed.
It's up to each chart to declare it's verify installation
method to perform checks.

Ticket: ENG-1342