feat(ENG-1910): let karpenter handle worker nodes k8s upgrade
Karpenter asynchronously manages worker node upgrades, meaning that when the control plane version is updated, it detects the change in the AMI (Amazon Machine Image) and identifies which nodes need upgrading by marking them as "drifted".
Once a new AMI is detected, Karpenter marks the existing version as drifted and starts upgrading it automatically.

Ticket: ENG-1910