feat(azure): remove docker images pull secrets
This PR allows to make use of the AKS cluster <> ACR registry linkage
provided by Azure allowing the cluster to pull images from the registry.
It removes the need to have a docker image pull secrets on pods which is
far from ideal especially since we will use Qovery service principal to
do so in a near future.