chore(reporter): Allow the deployment task to be FnMut
Terraform needs to be able to drop/mut some state between
the reporter and the task deployment/runner.

Today the interface was not allowing it because it was taking a Fn().
Alternative would have to use a mutex/interrior mutability instead of a
Option.
But as this may be useful in the future, extended the interface to support
FnMut